refactor(tools): split dispatch.rs into ToolEntry registry and per-tool builtin modules
