+++ crates/genie-core/src/tools/builtin/mod.rs
@@ -1,0 +1,16 @@
+pub mod action_history;
+pub mod calculate;
+pub mod get_time;
+pub mod get_weather;
+pub mod home_control;
+pub mod home_status;
+pub mod home_undo;
+pub mod memory_forget;
+pub mod memory_recall;
+pub mod memory_status;
+pub mod memory_store;
+pub mod play_media;
+pub mod set_timer;
+pub mod skill_runner;
+pub mod system_info;
+pub mod web_search;
