+++ crates/genie-core/src/tools/mod.rs
@@ -1,16 +1,24 @@
 pub mod actuation;
+pub mod builtin;
 pub mod calc;
+pub mod config;
 pub mod dispatch;
+pub mod dispatcher;
 mod home;
 pub mod parser;
 pub mod quick;
+mod registry;
 mod system;
 pub mod timer;
 mod weather;
 pub(crate) mod web_search;
 
 pub use actuation::{PendingConfirmation, RequestOrigin};
-pub use dispatch::{ToolActionClass, ToolCall, ToolDispatcher, ToolExecutionContext, ToolResult};
+pub use config::ToolDispatcherConfig;
+pub use dispatch::{
+    ToolActionClass, ToolCall, ToolDef, ToolEntry, ToolExecutionContext, ToolResult,
+};
+pub use dispatcher::ToolDispatcher;
 pub use parser::{
     UNPARSED_TOOL_CALL_FALLBACK, is_unparsed_tool_call, parse_tool_calls_for_eval, try_tool_call,
     try_tool_call_with_context,
